refactor(analyzer): split content.rs into separate modules for 400-line limit
Split the monolithic content.rs (948 lines) into smaller focused modules:
- config.rs (40 lines) - ExtractionConfig
- extractor.rs (179 lines) - ContentExtractor pipeline orchestration
- transforms/cleaner.rs (396 lines) - ContentCleaner state machine
- transforms/dedupe.rs (173 lines) - DeduplicateProgressLines
- transforms/normalize.rs (159 lines) - NormalizeWhitespace, FilterEmptyEvents

All files now comply with the 400-line maximum from coding-principles.md.
